Disease/requirements.txt
Flask==2.0.3
tensorflow==2.11.0
numpy
pillow

cd disease
python -m venv venv
.\venv\Scripts\activate